feat(DTFS2-6794): refactor GEF user fixtures
