+++ e2e-tests/gef/cypress/e2e/unissued-facilities/unissued-facilities-change-submit-ukef-MIN.spec.js
@@ -7,7 +7,7 @@ import CONSTANTS from '../../fixtures/constants';
 import dateConstants from '../../../../e2e-fixtures/dateConstants';
 
 import { MOCK_APPLICATION_MIN } from '../../fixtures/mocks/mock-deals';
-import { MOCK_USER_MAKER } from '../../fixtures/mocks/mock-user-maker';
+import { BANK1_MAKER1 as MOCK_USER_MAKER } from '../../fixtures/mocks/users';
 import {
   MOCK_FACILITY_ONE, MOCK_FACILITY_TWO, MOCK_FACILITY_THREE, MOCK_FACILITY_FOUR,
 } from '../../fixtures/mocks/mock-facilities';
