--- e2e-tests/gef/cypress/fixtures/mocks/mock-checker.js
@@ -1,21 +1,0 @@
-exports.MOCK_USER_CHECKER = {
-  _id: '123',
-  username: 'BANK1_MAKER1',
-  password: 'AbC!2345',
-  firstname: 'First',
-  surname: 'Last',
-  email: 'maker1@ukexportfinance.gov.uk',
-  timezone: 'Europe/London',
-  roles: ['checker'],
-  bank: {
-    id: '9',
-    name: 'UKEF test bank (Delegated)',
-    mga: ['mga_ukef_1.docx', 'mga_ukef_2.docx'],
-    emails: [
-      'maker1@ukexportfinance.gov.uk',
-      'checker1@ukexportfinance.gov.uk',
-    ],
-    companiesHouseNo: 'UKEF0001',
-    partyUrn: '00318345',
-  },
-};
