--- e2e-tests/gef/cypress/fixtures/mocks/mock-tfm-user.js
@@ -1,9 +1,0 @@
-exports.MOCK_TFM_USER = {
-  username: 'T1_USER_1',
-  email: 'test@testing.com',
-  password: 'AbC!2345',
-  teams: ['TEAM1'],
-  timezone: 'Europe/London',
-  firstName: 'Joe',
-  lastName: 'Bloggs',
-};
