+++ e2e-tests/gef/cypress/support/commands/submitDealAfterUkefIds.js
@@ -1,4 +1,4 @@
-const { MOCK_TFM_USER } = require('../../fixtures/mocks/mock-tfm-user');
+const { T1_USER_1 } = require('../../fixtures/mocks/users');
 const { tfmLogin, submitDealAfterUkefIds } = require('./api');
 
 /**
@@ -9,7 +9,7 @@ const { tfmLogin, submitDealAfterUkefIds } = require('./api');
  * @param {Object} checker
  */
 const submitDealAfterUkefIdsCall = (dealId, dealType, checker) => {
-  const { username, password } = MOCK_TFM_USER;
+  const { username, password } = T1_USER_1;
 
   tfmLogin(username, password).then((token) => submitDealAfterUkefIds(dealId, dealType, checker, token));
 };
